tools: resolve compiler warnings on x32
tools/freebusy.cpp: In function "void output_event(time_t, time_t, uint32_t, const char*, const char*, const char*, long int, long int, long int, long int, long int)":
tools/freebusy.cpp:729:42: warning: format "%lu" expects argument of type "long unsigned int", but argument 2 has type "time_t" {aka "long long int"} [-Wformat=]
  729 | printf("{\"StartTime\":%lu, ", start_time);
tools/freebusy.cpp:730:39: warning: format "%lu" expects argument of type "long unsigned int", but argument 2 has type "time_t" {aka "long long int"} [-Wformat=]
  730 | printf("\"EndTime\":%lu, ", end_time);
tools/genimport.cpp: In function "void gi_folder_map_write(const gi_folder_map_t&)":
tools/genimport.cpp:467:22: warning: comparison of integer expressions of different signedness: "int" and "uint32_t" {aka "unsigned int"} [-Wsign-compare]
  467 | else if (ret != ep.m_offset)
tools/genimport.cpp: In function "void gi_name_map_write(const gi_name_map&)":
tools/genimport.cpp:518:22: warning: comparison of integer expressions of different signedness: "int" and "uint32_t" {aka "unsigned int"} [-Wsign-compare]
  518 | else if (ret != ep.m_offset)
tools/timer.cpp: In function "void* tmr_thrwork(void*)":
tools/timer.cpp:526:62: warning: format "%ld" expects argument of type "long int", but argument 4 has type "time_t" {aka "long long int"} [-Wformat=]
  526 | temp_len = sprintf(temp_line, "%d\t%ld\t", ptimer->t_id,

+++ tools/freebusy.cpp
@@ -726,8 +726,8 @@ static void output_event(time_t start_time, time_t end_time,
 	char tmp_buff[4096];
 	
 	if (NULL == g_tz_component) {
-		printf("{\"StartTime\":%lu, ", start_time);
-		printf("\"EndTime\":%lu, ", end_time);
+		printf("{\"StartTime\":%lld, ", static_cast<long long>(start_time));
+		printf("\"EndTime\":%lld, ", static_cast<long long>(end_time));
 	} else {
 		ical_utc_to_datetime(g_tz_component, start_time, &itime);
 		printf("{\"StartTime\":\"%d-%02d-%02dT%02d:%02d:%02d\", ",
